requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
pdfplumber>=0.9.0
cryptography>=41.0.0
PyJWT>=2.8.0
gspread>=5.10.0
google-auth>=2.22.0